Handshake is stored in MySQL database
You have to setup the MySQL database first (see README file). When the client connects to the webserver for the first time,
an entry for the client's uuid is created in the database table. On each following update the entry is updated.

Warning:
If you don't have a secure and/or private environment someone may fill up your database table since each new value is simply added!